Re-order interface delete to hopefully be more reliable.
